alterando tipo de campo preco

+++ db/migrate/20150809232416_create_purchases.rb
@@ -2,7 +2,7 @@ class CreatePurchases < ActiveRecord::Migration
   def change
     create_table :purchases do |t|
       t.references :vendor, index: true
-      t.integer :total_price
+      t.decimal :total_price
 
       t.timestamps
     end
